Moved additional decision functionality into decisions.js file
