+++ gemp-module/gemp-stccg-client/src/main/web/js/gemp-022/gameUi.js
@@ -15,7 +15,6 @@ import strengthIconImg from "../../images/o_icon_strength.png";
 import vitalityIconImg from "../../images/o_icon_strength.png";
 import compassIconImg from "../../images/o_icon_strength.png";
 import resistanceIconImg from "../../images/o_icon_strength.png";
-import awaitingActionAudio from "../../src/assets/awaiting_decision.mp3";
 
 
 export default class GameTableUI {
@@ -1408,200 +1407,6 @@ export default class GameTableUI {
 
         return cardDiv;
     }
-
-    PlayAwaitActionSound() {
-        let audio = new Audio(awaitingActionAudio);
-        if(!document.hasFocus() || document.hidden || document.msHidden || document.webkitHidden)
-        {
-		    audio.play();
-        }
-    }
-
-    createActionChoiceContextMenu(actions, event, selectActionFunction) {
-        // Remove context menus that may be showing
-        $(".contextMenu").remove();
-
-        var div = $("<ul class='contextMenu'></ul>");
-        for (var i = 0; i < actions.length; i++) {
-            var action = actions[i];
-            var text = action.actionText;
-            div.append("<li><a href='#" + action.actionId + "'>" + text + "</a></li>");
-        }
-
-        $("#main").append(div);
-
-        var x = event.pageX;
-        var y = event.pageY;
-        $(div).css({left: x, top: y}).fadeIn(150);
-
-        $(div).find('A').mouseover(
-            function () {
-                $(div).find('LI.hover').removeClass('hover');
-                $(this).parent().addClass('hover');
-            }).mouseout(function () {
-            $(div).find('LI.hover').removeClass('hover');
-        });
-
-        var getRidOfContextMenu = function () {
-            $(div).remove();
-            $(document).unbind("click", getRidOfContextMenu);
-            return false;
-        };
-
-        // When items are selected
-        $(div).find('A').unbind('click');
-        $(div).find('LI:not(.disabled) A').click(function () {
-            $(document).unbind('click', getRidOfContextMenu);
-            $(".contextMenu").remove();
-
-            var actionId = $(this).attr('href').substr(1);
-            selectActionFunction(actionId);
-            return false;
-        });
-
-        // Hide bindings
-        setTimeout(function () { // Delay for Mozilla
-            $(document).click(getRidOfContextMenu);
-        }, 0);
-    }
-
-    recalculateCardSelectionOrder(cardArray) {
-        for (const [index, cardId] of cardArray.entries()) {
-            let divToChange = getCardDivFromId(cardId);
-            divToChange.attr("selectedOrder", index + 1); // use a 1-index
-        }
-    }
-
-    recalculateAllowedSelectionFromMaxCSS(cardIds, selectedCardIds, max) {
-        if (max === 0) {
-            console.error("Max is 0, setting all cards to not selectable. This is probably a server bug.");
-            for (const cardId of cardIds.values()) {
-                getCardDivFromId(cardId).removeClass("selectableCard").removeClass("selectedCard").addClass("notSelectableCard").removeClass("selectedBadge").removeAttr("selectedOrder");
-            }
-            return;
-        }
-        else {
-            for (const cardId of cardIds.values()) {
-                if (selectedCardIds.length === 0) {
-                    // everything is selectable
-                    getCardDivFromId(cardId).addClass("selectableCard").removeClass("selectedCard").removeClass("notSelectableCard").removeClass("selectedBadge").removeAttr("selectedOrder");
-                }
-                else {
-                    // selected
-                    if (selectedCardIds.includes(cardId)) {
-                        getCardDivFromId(cardId).removeClass("selectableCard").removeClass("notSelectableCard").addClass("selectedCard").addClass("selectedBadge");
-                    }
-                    // not selected
-                    else {
-                        // we hit the max, gray out unselected cards since we can't add more
-                        if (selectedCardIds.length === max) {
-                            getCardDivFromId(cardId).addClass("notSelectableCard").removeClass("selectableCard").removeClass("selectedCard").removeClass("selectedBadge").removeAttr("selectedOrder");
-                            continue;
-                        }
-                        else {
-                            getCardDivFromId(cardId).addClass("selectableCard").removeClass("selectedCard").removeClass("notSelectableCard").removeClass("selectedBadge").removeAttr("selectedOrder");
-                        }
-                    }
-                }
-            }
-            return;
-        }
-    }
-
-    recalculateAllowedCombinationsAndCSS(cardIds, selectedCardIds, jsonCombinations, max) {
-        if (typeof(jsonCombinations) !== 'object' || jsonCombinations == null) {
-            let inc_type = typeof(jsonCombinations);
-            throw new TypeError(`jsonCombinations must be an Object not a ${inc_type} or null.`);;
-        }
-
-        let allowedCombinationsRemaining = new Set();
-        
-        if (selectedCardIds.length === 0) {
-            // DEBUG: console.log("No selected cards.");
-            allowedCombinationsRemaining = new Set(cardIds);
-            // DEBUG: console.log(`Allowed combinations remaining: ${Array.from(allowedCombinationsRemaining)}`);
-        }
-        else if (selectedCardIds.length === 1) {
-            // selected one card
-            // DEBUG: console.log(`Selected cards: ${selectedCardIds}`);
-            const cardId = selectedCardIds[0];
-            let this_card_allowed = new Array();
-            // DEBUG: console.log("let_this_card_allowed created");
-            // DEBUG: console.log(jsonCombinations[cardId]);
-            for (const compatible_cardId of jsonCombinations[cardId]) {
-                // DEBUG: console.log("iterating through cards in jsonCombinations[" + cardId + "]");
-                this_card_allowed.push(compatible_cardId);
-            }
-            const this_allowed_as_set = new Set(this_card_allowed);
-            allowedCombinationsRemaining = this_allowed_as_set;
-            
-            // DEBUG: console.log(`Allowed combinations remaining: ${Array.from(allowedCombinationsRemaining)}`);
-        }
-        else {
-            // selected two or more cards
-            // DEBUG: console.log(`Selected cards: ${selectedCardIds}`);
-            for (const [index, cardId] of selectedCardIds.entries()) {
-                let this_card_allowed = new Array();
-                // DEBUG: console.log("let_this_card_allowed created");
-                // DEBUG: console.log(jsonCombinations[cardId]);
-                for (const compatible_cardId of jsonCombinations[cardId]) {
-                    // DEBUG: console.log("iterating through cards in jsonCombinations[" + cardId + "]");
-                    this_card_allowed.push(compatible_cardId);
-                }
-                const this_allowed_as_set = new Set(this_card_allowed);
-
-                if (index === 0) {
-                    // Don't use .intersection on the first pass, since the intersection of empty set and valid choices is nothing.
-                    allowedCombinationsRemaining = this_allowed_as_set;
-                }
-                else {
-                    allowedCombinationsRemaining = allowedCombinationsRemaining.intersection(this_allowed_as_set);
-                }
-                // DEBUG: console.log(`Allowed combinations remaining: ${Array.from(allowedCombinationsRemaining)}`);
-            }
-        }
-
-
-        // Apply CSS
-        // BUG: Normally I'd split this into another function but when I did, JQuery
-        //      didn't pass the Sets around properly. IDK. One big function it is.
-        for (const cardId of cardIds.values()) {
-            if (selectedCardIds.length === 0) {
-                // everything is selectable
-                // DEBUG: console.log("Everything is selectable.");
-                getCardDivFromId(cardId).addClass("selectableCard").removeClass("selectedCard").removeClass("notSelectableCard").removeClass("selectedBadge").removeAttr("selectedOrder");
-            }
-            else {
-                // selected
-                if (selectedCardIds.includes(cardId)) {
-                    getCardDivFromId(cardId).removeClass("selectableCard").removeClass("notSelectableCard").addClass("selectedCard").addClass("selectedBadge");
-                }
-                // not selected
-                else {
-                    // we hit the max, treat unselected cards as if they are not compatible
-                    if (selectedCardIds.length === max) {
-                        getCardDivFromId(cardId).addClass("notSelectableCard").removeClass("selectableCard").removeClass("selectedCard").removeClass("selectedBadge").removeAttr("selectedOrder");
-                        continue;
-                    }
-
-                    // Not selected, not at the max, and compatible with other selected cards
-                    if (allowedCombinationsRemaining.has(cardId)) {
-                        // DEBUG: console.log(`Not selected, compatible: ${cardId}`);
-                        getCardDivFromId(cardId).addClass("selectableCard").removeClass("selectedCard").removeClass("notSelectableCard").removeClass("selectedBadge").removeAttr("selectedOrder");
-                    }
-                    // Not selected, not at the max, but not compatible with other selected cards
-                    else {
-                        // DEBUG: console.log(`Not selected, not compatible: ${cardId}`);
-                        // same as above but w/o selectableCard
-                        getCardDivFromId(cardId).addClass("notSelectableCard").removeClass("selectableCard").removeClass("selectedCard").removeClass("selectedBadge").removeAttr("selectedOrder");
-                    }
-                }
-            }
-            
-        }
-        
-    }
-
 
     clearSelection() {
         $(".selectableCard").removeClass("selectableCard").data("action", null);
